+++ src/frontend/environments/environment.prod.ts
@@ -14,6 +14,7 @@ export const environment = {
     sdocActionTagWritable: false,
     allowAutoPlay: false,
     sdocMaxItemsPerAlbum: -1,
+    cookieLawSeenName: 'cookieLawSeenV20180525',
     trackingProviders: [Angulartics2Piwik]
 };